requirements.txt
inquirer
python-slugify